+++ stdlib_test.go
@@ -1,26 +1,17 @@
-package stdlib
+package stdlib_test
 
 import (
 	"errors"
 	"testing"
+
+	"github.com/goph/stdlib"
+	"github.com/goph/stdlib/internal/testing/assert"
 )
 
 func TestMust_Panics(t *testing.T) {
-	defer func() {
-		if r := recover(); r == nil {
-			t.Errorf("The code did not panic")
-		}
-	}()
-
-	Must(errors.New("should panic"))
+	assert.Panics(t, func() { stdlib.Must(errors.New("should panic")) })
 }
 
 func TestMust_DoesNotPanic(t *testing.T) {
-	defer func() {
-		if r := recover(); r != nil {
-			t.Errorf("The code should not panic")
-		}
-	}()
-
-	Must(nil)
+	assert.NotPanics(t, func() { stdlib.Must(nil) })
 }