+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - **internal/testing:** `Equal` assertion
+- **internal/testing:** Panic assertions
 
 
 ## 0.6.0 - 2017-07-06